# openapi.yaml
components:
...
    securitySchemes:
        anthropicGatewayKey:
            description: Gateway token on Anthropic-native ingress paths.
            in: header
            name: x-api-key
            type: apiKey
        bearerAuth:
            bearerFormat: API key
            description: 'Send Authorization: Bearer <token> when auth is enabled.'
            scheme: bearer
            type: http
        geminiGatewayKey:
            description: Gateway token on Gemini-native ingress paths.
            in: header
            name: x-goog-api-key
            type: apiKey
info:
    description: Compact provider-native LLM gateway with provider-aware routing, capability filtering, quota enforcement, SSE streaming, and provider-specific passthrough fields.
    title: llmgw
    version: v1
openapi: 3.0.3
paths:
    /docs:
        get:
            operationId: docsRedirect
            responses:
                "301":
                    description: Redirects to /docs/index.html.
            security:
                - bearerAuth: []
            summary: Redirect to Swagger UI
            tags:
                - docs
    /docs/index.html:
        get:
            operationId: docs
            responses:
                "200":
                    content:
                        text/html:
                            schema:
                                type: string
                    description: Swagger UI backed by /openapi.json.
            security:
                - bearerAuth: []
            summary: Swagger UI
            tags:
                - docs
...
    /openapi.json:
        get:
            operationId: openapiJSON
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                additionalProperties: true
                                type: object
                    description: Live OpenAPI document for the gateway surface.
            security:
                - bearerAuth: []
            summary: Generated OpenAPI JSON
            tags:
                - docs
    /openapi.yaml:
        get:
            operationId: openapiYAML
            responses:
                "200":
                    content:
                        application/yaml:
                            schema:
                                type: string
                        text/yaml:
                            schema:
                                type: string
                    description: Live OpenAPI document for the gateway surface.
            security:
                - bearerAuth: []
            summary: Generated OpenAPI YAML
            tags:
                - docs
...
    /v1/models:
        get:
            description: Returns only models allowed by the authenticated principal's model and provider ACLs.
            operationId: listModels
            responses:
                "200":
                    content:
                        application/json:
                            example:
                                data:
                                    - id: claude-sonnet-4-6
                                      object: model
                                      owned_by: llmgw
                                object: list
                            schema:
                                properties:
                                    data:
                                        items:
                                            properties:
                                                id:
                                                    type: string
                                                object:
                                                    type: string
                                                owned_by:
                                                    type: string
                                            type: object
                                        type: array
                                    object:
                                        type: string
                                type: object
                    description: Models visible to the caller.
                default:
                    content:
                        application/json:
                            example:
                                error:
                                    message: request failed
                                    type: invalid_request_error
                            schema:
                                properties:
                                    error:
                                        properties:
                                            code:
                                                type: string
                                            message:
                                                type: string
                                            param:
                                                type: string
                                            type:
                                                type: string
                                        type: object
                                type: object
                    description: Error response.
            security:
                - bearerAuth: []
            summary: List accessible route models
            tags:
                - system
...
servers:
    - description: Configured gateway listener
      url: /
...